+++ test/MapTests.cs
@@ -8,10 +8,8 @@ public sealed class MapTests
         await Assert.That(Option.Success(1).Map(i => i + 1)).IsSuccess(2);
         await Assert.That(Option.Success(1).Map(i => Option.Success(i))).IsSuccess(1);
         await Assert.That(Result.Success(1).Map(i => i + 1)).IsSuccess(2);
-        await Assert.That(Result.Success(1).Map(i => i + 1, e => e.Message)).IsSuccess(2);
         await Assert.That(Result.Success(1).Map(i => Result.Success(i))).IsSuccess(1);
         await Assert.That(Result.Success<int, string>(1).Map(i => i + 1)).IsSuccess(2);
-        await Assert.That(Result.Success<int, int>(1).Map(i => i + 1, error => error.ToString())).IsSuccess(2);
         await Assert.That(Result.Success<int, string>(5).Map(i => i + 1, error => new Exception(error))).IsSuccess();
         await Assert.That(Result.Success<int, string>(1).Map(i => Result.Success<int, string>(i))).IsSuccess(1);
         await Assert.That(RefOption.Success<Span<char>>([]).Map(s => new string(s))).IsSuccess("");
@@ -30,11 +28,9 @@ public sealed class MapTests
         await Assert.That(Option.Error<int>().Map(i => Option.Success(i))).IsError();
         await Assert.That(Option.Success(1).Map(i => Option.Error<int>())).IsError();
         await Assert.That(Result.Error<int>().Map(i => i + 1)).IsError();
-        await Assert.That(Result.Error<int>().Map(i => i + 1, e => e.Message)).IsError();
         await Assert.That(Result.Error<int>(new FormatException()).Map(i => Result.Success(i))).IsErrorOfType<int, FormatException>();
         await Assert.That(Result.Success(1).Map(i => Result.Error<int>(new ArgumentException()))).IsErrorOfType<int, ArgumentException>();
         await Assert.That(Result.Error<int, string>("").Map(i => i + 1)).IsError("");
-        await Assert.That(Result.Error<int, int>(5).Map(i => i + 1, error => error.ToString())).IsError("5");
         await Assert.That(Result.Error<int, string>("5").Map(i => i + 1, error => new Exception(error))).IsError();
         await Assert.That(Result.Error<int, string>("nay").Map(i => Result.Success<int, string>(i))).IsError("nay");
         await Assert.That(Result.Success<int, string>(1).Map(i => Result.Error<int, string>("nay"))).IsError("nay");