+++ changelog.md
@@ -6,4 +6,4 @@
 ## 0.2.2
 - moved nullable extensions to Ametrin.Optional.Numerics
 - updated TUnit to 18.9
-- removed obsolete (Try)Select use (Try)Map instead
+- removed obsolete `(Try)Select` use `(Try)Map` instead